requirements.txt
tablib

django
django-humanity
django-tablib